+++ project/libs/certbot/src/lib/certbot.ts
@@ -1,6 +1,7 @@
 import { spawnSync, SpawnSyncReturns } from 'child_process';
 
 import { getConfig, getLetsEncryptServer } from '@tx/config';
+import { Domain } from '@tx/domain';
 import { getEnv } from '@tx/environment';
 import { logger } from '@tx/logger';
 import { getStore } from '@tx/store';
@@ -9,21 +10,17 @@ import { getStore } from '@tx/store';
  * Request a certificate for the primary domain and optional domains.
  * Let certbot descide if an existing certificate needs to be updated.
  *
- * @param primaryDomain Primary domain for certificate
- * @param optionalDomains Domain variants to add to the same certificate
+ * @param domain Domains for certificate
  * @returns `true` when request was successful
- *
- * @todo
- * #TODO: Implement support for `optionalDomains`
  */
-export const requestCertificate = (
-	primaryDomain: string,
-	optionalDomains?: string[]
-): boolean => {
-	logger.info(`Request certificate for primary domain ${primaryDomain}...`);
+export const requestCertificate = (domain: Domain): boolean => {
+	logger.info(`Request certificate for primary domain ${domain.primary}...`);
+	if (domain.optional?.length) {
+		logger.info(`Apply domains to certificate: ${domain.optional.join(' ')}`);
+	}
 
 	// Optional domains are provided with `-d` flag before each domain
-	optionalDomains = optionalDomains ?? [];
+	const optionalDomains = domain.optional ?? [];
 	const includeDomainArg = `${
 		optionalDomains.length ? '-d ' : ''
 	}${optionalDomains.join(' -d ')}`;
@@ -53,7 +50,7 @@ export const requestCertificate = (
 		'--server',
 		`${getLetsEncryptServer()}`,
 		'--cert-name',
-		`${primaryDomain}`,
+		`${domain.primary}`,
 		`${includeDomainArg}`,
 		`${forceRenewal}`,
 		`${dryRun}`,
@@ -65,9 +62,9 @@ export const requestCertificate = (
 		// Spawn command syncron and hence request the certificate
 		status = spawnSync(command, args);
 	} else {
-		logger.info('Running in isolated mode, no request will be made!');
+		logger.info('<<< Running in isolated mode, no request will be made! >>>');
 		logger.info('certbot request command:');
-		logger.info(command);
+		logger.info(`${command} ${args.join(' ')}`);
 		status = { error: null } as SpawnSyncReturns<string>;
 	}
 	if (status.error) {