+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# 0.4 (2021-02-01)
+
+Features
+
+- Support multiple domains on a certificate with a common host
+- Send SIGUSR2 to trigger force renewal of certificates
+
+Enhancements
+
+- Isolated test terminates after the first loop..
+- Analyze Nginx captured error to see if it's a warning and log accordingly
+- Apply gzip compression via `conf.d/gzip.conf`
+
+Misc
+
+- Add to readme how to revoke certificate using `certbot` inside container
+- Much more documentation added to readme
+
 # 0.3 (2021-01-24)
 
 Features